refactor: Edit base viewmodel
