--- app/src/main/java/org/sopt/xmlStudy/core/base/SideEffect.kt
@@ -1,3 +1,0 @@
-package org.sopt.xmlStudy.core.base
-
-interface SideEffect
