+++ app/src/main/java/org/sopt/xmlStudy/core/base/ui/UiState.kt
@@ -1,0 +1,3 @@
+package org.sopt.xmlStudy.core.base.ui
+
+interface UiState
